+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# vX.Y.Z (yyyy-mm-dd)
+
+- **Breaking change:** don't copy source .scss/.sass files to build directory by default.
+  Use `copy-sources: true` to keep the old behavior.
+
 # v0.2.2 (2014-05-23)
 
 - Require Barback 0.13.x.